fix: prefer module build-time tpc
