+++ scripts/utils.ts
@@ -1,0 +1,108 @@
+import type { ExternalScript, Script } from 'third-party-capital'
+import { genImport } from 'knitwork'
+import type { HeadEntryOptions } from '@unhead/vue'
+import type { TpcDescriptor } from './generateTpcScripts'
+
+const HEAD_VAR = '__head'
+const INJECTHEAD_CODE = `const ${HEAD_VAR} =  injectHead()`
+
+export async function generateTpcContent(input: TpcDescriptor) {
+  if (!input.tpcData.scripts)
+    throw new Error('input.data has no scripts !')
+
+  // replace all empty spaces with nothing
+  const titleKey = input.label.replace(/\s/g, '')
+
+  const mainScript = input.tpcData.scripts?.find(({ key }) => key === input.tpcKey) as ExternalScript
+
+  if (!mainScript)
+    throw new Error(`no main script found for ${input.tpcKey} in third-party-capital`)
+
+  const mainScriptOptions = getScriptInputOption(mainScript)
+
+  const imports = new Set<string>([
+    'import { withQuery } from \'ufo\'',
+    'import { useRegistryScript } from \'#nuxt-scripts-utils\'',
+    'import type { RegistryScriptInput } from \'#nuxt-scripts\'',
+    `import type { ${input.tpcTypeImport} } from 'third-party-capital'`,
+  ])
+
+  const chunks: string[] = []
+  const functionBody: string[] = []
+
+  const hasParams = mainScript.params?.length
+
+  if (hasParams) {
+    imports.add(genImport('#nuxt-scripts-validator', ['object', 'string']))
+    // need schema validation from tpc
+    chunks.push(`export const ${titleKey}Options = object({${mainScript.params?.map(p => `${p}:  string()`)}})`)
+  }
+
+  chunks.push(`
+declare global {
+  interface Window extends ${input.tpcTypeImport} {}
+}`)
+
+  const clientInitCode: string[] = []
+
+  if (input.tpcData.stylesheets) {
+    if (!functionBody.includes(INJECTHEAD_CODE)) {
+      functionBody.unshift(INJECTHEAD_CODE)
+    }
+    functionBody.push(`${HEAD_VAR}.push({link: ${JSON.stringify(input.tpcData.stylesheets.map(s => ({ ref: 'stylesheet', href: s })))}})`)
+  }
+
+  for (const script of input.tpcData.scripts) {
+    if ('code' in script)
+      clientInitCode.push(replaceTokenToRuntime(script.code))
+
+    if (script === mainScript)
+      continue
+
+    if ('url' in script) {
+      functionBody.push(`${HEAD_VAR}.push({scripts:{ async: true, src: ${script.url} }},${JSON.stringify(getScriptInputOption(script))})`)
+    }
+  }
+
+  chunks.push(`export type ${titleKey}Input = RegistryScriptInput${hasParams ? `<typeof ${titleKey}Options>` : ''}`)
+
+  chunks.push(`
+export function ${input.registry.import!.name}<T extends ${input.tpcTypeImport}>(_options?: ${titleKey}Input) {
+${functionBody.join('\n')}
+  return useRegistryScript${hasParams ? `<T, typeof ${titleKey}Options>` : ''}(_options?.key || '${input.key}', options => ({
+        scriptInput: {
+            src: withQuery('${mainScript.url}', {${mainScript.params?.map(p => `${p}: options?.${p}`)}})
+        },
+        schema: import.meta.dev ? undefined : ${titleKey}Options,
+        scriptOptions: {
+            use: ${input.options.scriptOptions!.use!.toString()},
+            stub: import.meta.client ? undefined :  ${input.options.scriptOptions!.stub!.toString()},
+            ${input.options.scriptOptions?.performanceMarkFeature ? `performanceMarkFeature: ${JSON.stringify(input.options.scriptOptions?.performanceMarkFeature)},` : ''}
+            ${mainScriptOptions ? `...(${JSON.stringify(mainScriptOptions)})` : ''}
+        },
+        // eslint-disable-next-line
+        ${clientInitCode.length ? `clientInit: import.meta.server ? undefined : () => {${clientInitCode.join('\n')}},` : ''}
+    }), _options)
+}`)
+
+  chunks.unshift(...Array.from(imports))
+  chunks.unshift('// WARNING: This file is automatically generated, do not manually modify.')
+  return chunks.join('\n')
+}
+
+function replaceTokenToRuntime(code: string) {
+  return code.split(';').map(c => c.replaceAll(/'?\{\{(.*?)\}\}'?/g, 'options.$1')).join(';')
+}
+
+function getScriptInputOption(script: Script): HeadEntryOptions | undefined {
+  if (script.location === 'body') {
+    if (script.action === 'append') {
+      return { tagPosition: 'bodyClose' }
+    }
+    return { tagPosition: 'bodyOpen' }
+  }
+
+  if (script.action === 'append') {
+    return { tagPriority: 1 }
+  }
+}
