+++ src/registry.ts
@@ -1,4 +1,8 @@
 import { joinURL, withBase, withQuery } from 'ufo'
+import {
+  GoogleAnalytics,
+  GoogleTagManager,
+} from 'third-party-capital'
 import type { HotjarInput } from './runtime/registry/hotjar'
 import type { IntercomInput } from './runtime/registry/intercom'
 import type { SegmentInput } from './runtime/registry/segment'
@@ -226,6 +230,42 @@ export const registry: (resolve?: (s: string) => string) => RegistryScripts = (r
         name: 'useScriptNpm',
         // key is based on package name
         from: resolve('./runtime/registry/npm'),
+      },
+    },
+    {
+      label: 'Google Tag Manager',
+      category: 'tracking',
+      import: {
+        name: 'useScriptGoogleTagManager',
+        from: resolve('./runtime/registry/google-tag-manager'),
+      },
+      logo: `<svg xmlns="http://www.w3.org/2000/svg" width="32" height="32" viewBox="0 0 256 256"><path fill="#8AB4F8" d="m150.262 245.516l-44.437-43.331l95.433-97.454l46.007 45.091z"/><path fill="#4285F4" d="M150.45 53.938L106.176 8.731L9.36 104.629c-12.48 12.48-12.48 32.713 0 45.207l95.36 95.986l45.09-42.182l-72.654-76.407z"/><path fill="#8AB4F8" d="m246.625 105.37l-96-96c-12.494-12.494-32.756-12.494-45.25 0c-12.495 12.495-12.495 32.757 0 45.252l96 96c12.494 12.494 32.756 12.494 45.25 0c12.495-12.495 12.495-32.757 0-45.251"/><circle cx="127.265" cy="224.731" r="31.273" fill="#246FDB"/></svg>`,
+      scriptBundling(options) {
+        const data = GoogleTagManager(options)
+        const mainScript = data.scripts?.find(({ key }) => key === 'gtag')
+
+        if (mainScript && 'url' in mainScript && mainScript.url)
+          return mainScript.url
+
+        return false
+      },
+    },
+    {
+      label: 'Google Analytics',
+      category: 'tracking',
+      import: {
+        name: 'useScriptGoogleAnalytics',
+        from: resolve('./runtime/registry/google-analytics'),
+      },
+      logo: `<svg xmlns="http://www.w3.org/2000/svg" width="32" height="32" viewBox="0 0 256 256"><path fill="#8AB4F8" d="m150.262 245.516l-44.437-43.331l95.433-97.454l46.007 45.091z"/><path fill="#4285F4" d="M150.45 53.938L106.176 8.731L9.36 104.629c-12.48 12.48-12.48 32.713 0 45.207l95.36 95.986l45.09-42.182l-72.654-76.407z"/><path fill="#8AB4F8" d="m246.625 105.37l-96-96c-12.494-12.494-32.756-12.494-45.25 0c-12.495 12.495-12.495 32.757 0 45.252l96 96c12.494 12.494 32.756 12.494 45.25 0c12.495-12.495 12.495-32.757 0-45.251"/><circle cx="127.265" cy="224.731" r="31.273" fill="#246FDB"/></svg>`,
+      scriptBundling(options) {
+        const data = GoogleAnalytics(options)
+        const mainScript = data.scripts?.find(({ key }) => key === 'gtag')
+
+        if (mainScript && 'url' in mainScript && mainScript.url)
+          return mainScript.url
+
+        return false
       },
     },
   ]
